requirements.txt
arrow==1.2.3
async-timeout==4.0.3
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
Redis-Sentinel-Url==1.0.1
requests==2.31.0
rq==1.15.1
rq-dashboard==0.6.1
rq-dashboard-compatible==0.1
six==1.16.0
soupsieve==2.4.1
urllib3==2.0.4
Werkzeug==2.3.6
